third_party/WebKit/LayoutTests/svg/custom/get-text-element-transform-crash-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
CONSOLE ERROR: line 5: Uncaught TypeError: Failed to execute 'getTransformToElement' on 'SVGGraphicsElement': 1 argument required, but only 0 present.
 PASS
